small changes to collect bench results

+++ ld.shared/shared.cu
@@ -141,7 +141,7 @@ int main() {
 
   thread_args.flag = 1;
   // Launch kernel on the GPU
-  benchmark_alt<int><<<NUM_BLOCKS, THREADS_PER_BLOCK>>>(
+  benchmark_alt<long><<<NUM_BLOCKS, THREADS_PER_BLOCK>>>(
       d_X, d_startClk, d_stopClk, d_timeStart, d_timeStop);
 
   // Wait for GPU to finish
@@ -174,7 +174,7 @@ int main() {
 
   total_time = total_time / 1e9;
 
-  long bytes = sizeof(int) * 2 * ITERATIONS * THREADS_PER_BLOCK *
+  long bytes = sizeof(long) * 2 * ITERATIONS * THREADS_PER_BLOCK *
                NUM_BLOCKS;  // 2 for read and write
 
   // uint64_t fma =
